+++ src/main/java/me/m56738/easyarmorstands/config/EditorConfig.java
@@ -14,4 +14,5 @@ public class EditorConfig {
     public boolean allowEntities;
     public boolean allowPlayers;
     public boolean flattenArmorStands;
+    public boolean centeredPivot;
 }